pyproject.toml
[tool.poetry]
name = "elevation-tiler"
version = "0.1.0"
description = ""
authors = ["Daven Quinn <code@davenquinn.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
httpx = "^0.27.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
rio-rgbify = "^0.4.0"
morecantile = "^5.3.0"
rio-tiler = "^6.6.1"
pytest = "^8.2.1"
rio-cogeo = "^5.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
